Remove non-strict DOM load
Non-strict DOM load makes codes very complicated, and it requires many
accessors for pre-interpreted data to return `Option` or `Result`.
Additionally, it is hard to decide what is critical and what is not.

If users should allow broken or unexpected data, they should write
or use another DOM crate for their needs.