Enable iterator returned by `NodeHandle::children_by_name` to live longer
Previously, `self: &NodeHandle` is moved into `FnMut` closure, and this
shortened the lifetime of the returned iterator.
This is because the closure lives longer as the iterator, and it
requires `self: &NodeHandle` to be valid (i.e. `NodeHandle` object to be
alive).

New implementation generates `self.children()` iterator in `FnOnce`
closure, and it is immediately executed when `.map()` is executed.
Therefore, any iterators does not borrow `self`, and the result iterator
can live longer as `Tree`.